refactoring the onit cluster
