+++ pkg/k8s/cli/add.go
@@ -17,7 +17,8 @@ package cli
 import (
 	"fmt"
 
-	"github.com/onosproject/onos-test/pkg/onit"
+	"github.com/onosproject/onos-test/pkg/k8s"
+
 	"github.com/spf13/cobra"
 )
 
@@ -61,7 +62,7 @@ func getAddNetworkCommand() *cobra.Command {
 			configName, _ := cmd.Flags().GetString("preset")
 
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -80,7 +81,7 @@ func getAddNetworkCommand() *cobra.Command {
 
 			// Create the network configuration
 
-			config := &onit.NetworkConfig{
+			config := &k8s.NetworkConfig{
 				Config: configName,
 			}
 			if len(args) > 1 {
@@ -88,7 +89,7 @@ func getAddNetworkCommand() *cobra.Command {
 			}
 
 			// Update number of devices in the network configuration
-			onit.ParseMininetOptions(config)
+			k8s.ParseMininetOptions(config)
 
 			if err != nil {
 				exitError(err)
@@ -129,8 +130,8 @@ func getAddSimulatorCommand() *cobra.Command {
 			// Create the simulator configuration from the configured preset
 			configName, _ := cmd.Flags().GetString("preset")
 
-			// Get the onit controller
-			controller, err := onit.NewController()
+			// Get the k8s controller
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -148,7 +149,7 @@ func getAddSimulatorCommand() *cobra.Command {
 			}
 
 			// Create the simulator configuration
-			config := &onit.SimulatorConfig{
+			config := &k8s.SimulatorConfig{
 				Config: configName,
 			}
