+++ pkg/k8s/cli/create.go
@@ -17,7 +17,8 @@ package cli
 import (
 	"fmt"
 
-	"github.com/onosproject/onos-test/pkg/onit"
+	"github.com/onosproject/onos-test/pkg/k8s"
+
 	"github.com/spf13/cobra"
 )
 
@@ -52,25 +53,25 @@ func getCreateCommand() *cobra.Command {
 
 func initImageTags(imageTags map[string]string) {
 	if imageTags["config"] == "" {
-		imageTags["config"] = string(onit.Debug)
+		imageTags["config"] = string(k8s.Debug)
 	}
 	if imageTags["topo"] == "" {
-		imageTags["topo"] = string(onit.Debug)
+		imageTags["topo"] = string(k8s.Debug)
 	}
 	if imageTags["atomix"] == "" {
-		imageTags["atomix"] = string(onit.Latest)
+		imageTags["atomix"] = string(k8s.Latest)
 	}
 	if imageTags["raft"] == "" {
-		imageTags["raft"] = string(onit.Latest)
+		imageTags["raft"] = string(k8s.Latest)
 	}
 	if imageTags["simulator"] == "" {
-		imageTags["simulator"] = string(onit.Latest)
+		imageTags["simulator"] = string(k8s.Latest)
 	}
 	if imageTags["stratum"] == "" {
-		imageTags["stratum"] = string(onit.Latest)
+		imageTags["stratum"] = string(k8s.Latest)
 	}
 	if imageTags["test"] == "" {
-		imageTags["test"] = string(onit.Latest)
+		imageTags["test"] = string(k8s.Latest)
 	}
 
 }
@@ -92,8 +93,8 @@ func getCreateClusterCommand() *cobra.Command {
 
 			initImageTags(imageTags)
 
-			// Get the onit controller
-			controller, err := onit.NewController()
+			// Get the k8s controller
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -107,7 +108,7 @@ func getCreateClusterCommand() *cobra.Command {
 			}
 
 			// Create the cluster configuration
-			config := &onit.ClusterConfig{
+			config := &k8s.ClusterConfig{
 				Registry:      dockerRegistry,
 				Preset:        configName,
 				ImageTags:     imageTags,
@@ -139,13 +140,13 @@ func getCreateClusterCommand() *cobra.Command {
 	}
 
 	imageTags := make(map[string]string)
-	imageTags["config"] = string(onit.Debug)
-	imageTags["topo"] = string(onit.Debug)
-	imageTags["simulator"] = string(onit.Latest)
-	imageTags["stratum"] = string(onit.Latest)
-	imageTags["test"] = string(onit.Latest)
-	imageTags["atomix"] = string(onit.Latest)
-	imageTags["raft"] = string(onit.Latest)
+	imageTags["config"] = string(k8s.Debug)
+	imageTags["topo"] = string(k8s.Debug)
+	imageTags["simulator"] = string(k8s.Latest)
+	imageTags["stratum"] = string(k8s.Latest)
+	imageTags["test"] = string(k8s.Latest)
+	imageTags["atomix"] = string(k8s.Latest)
+	imageTags["raft"] = string(k8s.Latest)
 
 	cmd.Flags().StringP("config", "c", "default", "test cluster configuration")
 	cmd.Flags().String("docker-registry", "", "an optional host:port for a private Docker registry")
