+++ pkg/k8s/cli/debug.go
@@ -18,7 +18,8 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/onosproject/onos-test/pkg/onit"
+	"github.com/onosproject/onos-test/pkg/k8s"
+
 	"github.com/spf13/cobra"
 )
 
@@ -39,8 +40,8 @@ func getDebugCommand() *cobra.Command {
 		Example: debugExample,
 		Args:    cobra.MaximumNArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {
-			// Get the onit controller
-			controller, err := onit.NewController()
+			// Get the k8s controller
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -70,13 +71,13 @@ func getDebugCommand() *cobra.Command {
 				wg.Add(n)
 
 				asyncErrors := make(chan error, n)
-				freePorts, err := onit.GetFreePorts(n)
+				freePorts, err := k8s.GetFreePorts(n)
 				if err != nil {
 					exitError(err)
 				}
 
 				for index := range nodes {
-					go func(node onit.NodeInfo, port int) {
+					go func(node k8s.NodeInfo, port int) {
 						fmt.Println("Start Debugger for:", node.ID)
 						err = cluster.PortForward(node.ID, port, 40000)
 						asyncErrors <- err
@@ -110,6 +111,6 @@ func getDebugCommand() *cobra.Command {
 	cmd.Flags().Lookup("cluster").Annotations = map[string][]string{
 		cobra.BashCompCustom: {"__onit_get_clusters"},
 	}
-	cmd.Flags().IntP("port", "p", onit.DebugPort, "the local port to forward to the given resource")
+	cmd.Flags().IntP("port", "p", k8s.DebugPort, "the local port to forward to the given resource")
 	return cmd
 }
