+++ pkg/k8s/cli/run.go
@@ -19,9 +19,10 @@ import (
 	"os"
 	"time"
 
+	"github.com/onosproject/onos-test/pkg/k8s"
+
 	"github.com/onosproject/onos-test/test"
 
-	"github.com/onosproject/onos-test/pkg/onit"
 	"github.com/onosproject/onos-test/pkg/runner"
 	"github.com/spf13/cobra"
 )
@@ -164,8 +165,8 @@ func getRunBenchSuiteCommand(registry *runner.TestRegistry) *cobra.Command {
 }
 
 func runTestsRemote(cmd *cobra.Command, testID string, commandType string, tests []string, count int) {
-	// Get the onit controller
-	controller, err := onit.NewController()
+	// Get the k8s controller
+	controller, err := k8s.NewController()
 	if err != nil {
 		exitError(err)
 	}
