+++ pkg/k8s/console/spinner.go
