+++ pkg/k8s/rbac.go
@@ -1,0 +1,153 @@
+// Copyright 2019-present Open Networking Foundation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package k8s
+
+import (
+	corev1 "k8s.io/api/core/v1"
+	rbacv1 "k8s.io/api/rbac/v1"
+	k8serrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+// setupRBAC sets up role based access controls for the cluster
+func (c *ClusterController) setupRBAC() error {
+	if err := c.createClusterRole(); err != nil {
+		return err
+	}
+	if err := c.createClusterRoleBinding(); err != nil {
+		return err
+	}
+	if err := c.createServiceAccount(); err != nil {
+		return err
+	}
+	return nil
+}
+
+// createClusterRole creates the ClusterRole required by the Atomix controller and tests if not yet created
+func (c *ClusterController) createClusterRole() error {
+	role := &rbacv1.ClusterRole{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      c.clusterID,
+			Namespace: c.clusterID,
+		},
+		Rules: []rbacv1.PolicyRule{
+			{
+				APIGroups: []string{
+					"",
+				},
+				Resources: []string{
+					"pods",
+					"services",
+					"endpoints",
+					"persistentvolumeclaims",
+					"events",
+					"configmaps",
+					"secrets",
+				},
+				Verbs: []string{
+					"*",
+				},
+			},
+			{
+				APIGroups: []string{
+					"",
+				},
+				Resources: []string{
+					"namespaces",
+				},
+				Verbs: []string{
+					"get",
+				},
+			},
+			{
+				APIGroups: []string{
+					"apps",
+				},
+				Resources: []string{
+					"deployments",
+					"daemonsets",
+					"replicasets",
+					"statefulsets",
+				},
+				Verbs: []string{
+					"*",
+				},
+			},
+			{
+				APIGroups: []string{
+					"policy",
+				},
+				Resources: []string{
+					"poddisruptionbudgets",
+				},
+				Verbs: []string{
+					"*",
+				},
+			},
+			{
+				APIGroups: []string{
+					"k8s.atomix.io",
+				},
+				Resources: []string{
+					"*",
+				},
+				Verbs: []string{
+					"*",
+				},
+			},
+		},
+	}
+	_, err := c.kubeclient.RbacV1().ClusterRoles().Create(role)
+	if err != nil && !k8serrors.IsAlreadyExists(err) {
+		return err
+	}
+	return nil
+}
+
+// createClusterRoleBinding creates the ClusterRoleBinding required by the Atomix controller and tests for the test namespace
+func (c *ClusterController) createClusterRoleBinding() error {
+	roleBinding := &rbacv1.ClusterRoleBinding{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      c.clusterID,
+			Namespace: c.clusterID,
+		},
+		Subjects: []rbacv1.Subject{
+			{
+				Kind:      "ServiceAccount",
+				Name:      c.clusterID,
+				Namespace: c.clusterID,
+			},
+		},
+		RoleRef: rbacv1.RoleRef{
+			Kind:     "ClusterRole",
+			Name:     c.clusterID,
+			APIGroup: "rbac.authorization.k8s.io",
+		},
+	}
+	_, err := c.kubeclient.RbacV1().ClusterRoleBindings().Create(roleBinding)
+	return err
+}
+
+// createServiceAccount creates a ServiceAccount used by the Atomix controller
+func (c *ClusterController) createServiceAccount() error {
+	serviceAccount := &corev1.ServiceAccount{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      c.clusterID,
+			Namespace: c.clusterID,
+		},
+	}
+	_, err := c.kubeclient.CoreV1().ServiceAccounts(c.clusterID).Create(serviceAccount)
+	return err
+}
